backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
qdrant-client==1.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sentence-transformers==2.2.2
